services/scraper/pyproject.toml
[project]
name = "historicam-data-collection"
version = "0.1.0"
description = "Data collection scripts for HistoriCam"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "google-cloud-storage>=2.10.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "imagehash>=4.3.1",
    "bs4>=0.0.2",
]

# gcs_utils is mounted at runtime via volume mount
# Install at runtime with: uv pip install -e /app/lib/gcs_utils

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
